refactor, feat: config 패키지 구성 및 edgeWeightStrategy 컨버터 추가
