+++ src/main/java/nextstep/config/WebConfig.java
@@ -1,0 +1,23 @@
+package nextstep.config;
+
+import nextstep.config.converter.StringToEdgeWeightStrategyConverter;
+import org.springframework.context.ApplicationContext;
+import org.springframework.context.annotation.Configuration;
+import org.springframework.format.FormatterRegistry;
+import org.springframework.web.servlet.config.annotation.WebMvcConfigurer;
+
+@Configuration
+public class WebConfig implements WebMvcConfigurer {
+
+    private final ApplicationContext applicationContext;
+
+    public WebConfig(ApplicationContext applicationContext) {
+        this.applicationContext = applicationContext;
+    }
+
+    @Override
+    public void addFormatters(FormatterRegistry registry) {
+        registry.addConverter(new StringToEdgeWeightStrategyConverter(applicationContext));
+    }
+
+}
